separate out DB2 instrumentation with expensive matcher from the rest of JDBC
