+++ dd-java-agent/instrumentation/jdbc/src/main/java/datadog/trace/instrumentation/jdbc/AbstractConnectionInstrumentation.java
@@ -1,0 +1,51 @@
+package datadog.trace.instrumentation.jdbc;
+
+import static datadog.trace.agent.tooling.bytebuddy.matcher.DDElementMatchers.hasInterface;
+import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.nameStartsWith;
+import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
+import static java.util.Collections.singletonMap;
+import static net.bytebuddy.matcher.ElementMatchers.returns;
+import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
+
+import datadog.trace.agent.tooling.Instrumenter;
+import datadog.trace.bootstrap.ContextStore;
+import datadog.trace.bootstrap.InstrumentationContext;
+import datadog.trace.bootstrap.instrumentation.jdbc.DBQueryInfo;
+import java.sql.PreparedStatement;
+import java.util.Map;
+import net.bytebuddy.asm.Advice;
+import net.bytebuddy.description.method.MethodDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+
+public abstract class AbstractConnectionInstrumentation extends Instrumenter.Tracing {
+  public AbstractConnectionInstrumentation(String instrumentationName, String... additionalNames) {
+    super(instrumentationName, additionalNames);
+  }
+
+  @Override
+  public Map<String, String> contextStore() {
+    return singletonMap("java.sql.PreparedStatement", DBQueryInfo.class.getName());
+  }
+
+  @Override
+  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+    return singletonMap(
+        nameStartsWith("prepare")
+            .and(takesArgument(0, String.class))
+            // Also include CallableStatement, which is a sub type of PreparedStatement
+            .and(returns(hasInterface(named("java.sql.PreparedStatement")))),
+        AbstractConnectionInstrumentation.class.getName() + "$ConnectionPrepareAdvice");
+  }
+
+  public static class ConnectionPrepareAdvice {
+    @Advice.OnMethodExit(suppress = Throwable.class)
+    public static void addDBInfo(
+        @Advice.Argument(0) final String sql, @Advice.Return final PreparedStatement statement) {
+      ContextStore<PreparedStatement, DBQueryInfo> contextStore =
+          InstrumentationContext.get(PreparedStatement.class, DBQueryInfo.class);
+      if (null == contextStore.get(statement)) {
+        contextStore.putIfAbsent(statement, DBQueryInfo.ofPreparedStatement(sql));
+      }
+    }
+  }
+}
